Move transaction encoding tests to primitives/tests/encoding.rs
This change moves transaction consensus encoding and decoding tests,
along with shared helpers and constants, into a dedicated test file.
Removes the FIXME comment that requested this change.

The assert_eq! on the no-witnesses error is replaced with matches!
on the outer error type due to visibility constraints.
